+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 ### Changelog
 
+- **dev**
+  - Fix: Delay is wokring for "Switch to a tab when cursor is hovered on it"
 
 - **0.1.0**
   - Feature: Navigate to last page by clicking LMB while holding RMB (default: on)
   - Feature: Switch tabs by scroll wheel when hover on tab bar
   - Feature: Switch tabs by scroll wheel when holding right mouse button
-  - Feature: Switch to a tab when pointer is hovered on it
+  - Feature: Switch to a tab when cursor is hovered on it
   - Options can be changed inside Addons Manager